refactor: 리턴값을 소숫점 1자리까지로 반환

+++ src/main/java/rehabilitation/device/model/dto/FlexResponse.java
@@ -8,15 +8,15 @@ import java.time.LocalDateTime;
 @Getter
 public class FlexResponse {
 	private Long id;
-	private Integer thumb;
-	private Integer indexFinger;
-	private Integer middleFinger;
-	private Integer ringFinger;
-	private Integer littleFinger;
+	private double thumb;
+	private double indexFinger;
+	private double middleFinger;
+	private double ringFinger;
+	private double littleFinger;
 	private LocalDateTime date;
 
 	@Builder
-	public FlexResponse(Long id, Integer thumb, Integer indexFinger, Integer middleFinger, Integer ringFinger, Integer littleFinger, LocalDateTime date) {
+	public FlexResponse(Long id, double thumb, double indexFinger, double middleFinger, double ringFinger, double littleFinger, LocalDateTime date) {
 		this.id = id;
 		this.thumb = thumb;
 		this.indexFinger = indexFinger;
